2012-10-03 Roman - SYS: Function TSYS::sysTm() is added for allow fast access to time from periodic updated global memory cell.     Used for DAQ.JavaLikeCalc execution time limit control. - !!:DAQ.JavaLikeCalc: Time limit processing is made more optimal for time, from 5% to 12% JavaLike code execution faster.     Exception generation enabled on execution time limit exceed. - FIX:Archive: Single value, typical for 1h, request is fixed. - FIX:UI.Vision, UI.WebVision: Periods values processing for hours more is fixed.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@1926 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/ui/Vision/vis_shapes.cpp
@@ -2587,7 +2587,7 @@ void ShapeDiagram::TrendObj::loadTrendsData( bool full )
 	    val_tp  = atoi(req.attr("vtp").c_str());
 	    arh_beg = atoll(req.attr("beg").c_str());
 	    arh_end = atoll(req.attr("end").c_str());
-	    arh_per = atoi(req.attr("per").c_str());
+	    arh_per = atoll(req.attr("per").c_str());
 	}
     }
