+++ src/moduls/ui/WebVision/vca_sess.h
@@ -360,7 +360,7 @@ class VCADiagram : public VCAObj
 		int		mColor;		//Values line color
 		char		mWidth;		//Line width
 		//> Archive
-		int		arh_per;	//Archive period
+		int64_t		arh_per;	//Archive period
 		int64_t		arh_beg;	//Archive begin time
 		int64_t		arh_end;	//Archive end time
 		//> Values
